Move util code to parse package
