+++ parse/parse.go
@@ -1,11 +1,11 @@
-package main
+package parse
 
 import (
 	"log"
 	"strconv"
 )
 
-func intMustParse(str string) int {
+func GetIntOrFail(str string) int {
 	result, err := strconv.Atoi(str)
 	if err != nil {
 		log.Fatalf("\"%v\" is invalid integer\n", str)
